+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Adds `nova.trame.get_app` to make it easier to access `ThemedApp` methods from child components (thanks to John Duggan).
 * `InputField` now supports `type="button"` for creating buttons (thanks to John Duggan).
 * `InputField` automatically visually indicates required fields in Pydantic (thanks to John Duggan).
+* Children of `GridLayout`, `HBoxLayout`, and `VBoxLayout` can now override the `halign`, `valign`, and `stretch` parameters (thanks to John Duggan).
 * `DataSelector` filtering has been moved from a dialog to a field above the file list (thanks to John Duggan).
 
 ### nova-trame, 1.7.1